+++ src/spec4/layouts/__init__.py
@@ -1,171 +1,45 @@
 from __future__ import annotations
 
 import pathlib
-import re
 from typing import Any
 
 from dash import dcc, html
 import dash_mantine_components as dmc
 
-from spec4 import providers
-from spec4.app_constants import (
-    STATE_PHASES_COMPLETE,
-    STATE_REVIEW_COMPLETE,
-    STATE_STACK_COMPLETE,
-    STATE_VISION_COMPLETE,
+from spec4.layouts._chat import _agent_status_bar, _chat_action_buttons, _chat_layout
+from spec4.layouts._done import _done_layout
+from spec4.layouts._setup import _setup_layout
+from spec4.layouts._shared import (
+    _card,
+    _error,
+    _feature_card,
+    _footer,
+    _nav_drawer,
+    _render_message,
+    _reformat_inline_lists,
 )
 
-
-# ---------------------------------------------------------------------------
-# Shared UI helpers
-# ---------------------------------------------------------------------------
-
-
-def _card(*children: Any, **kwargs: Any) -> Any:
-    return dmc.Paper(list(children), p="md", radius="md", withBorder=True, **kwargs)
-
-
-def _feature_card(*children: Any, **kwargs: Any) -> Any:
-    return dmc.Paper(
-        list(children),
-        p="lg",
-        radius="lg",
-        withBorder=True,
-        className="feature-card",
-        **kwargs,
-    )
-
-
-def _error(msg: str) -> Any:
-    return dmc.Alert(msg, color="red", variant="light", mt="sm")
-
-
-# Navigation links shared between the footer and nav drawer.
-_NAV_LINKS: list[tuple[str, list[tuple[str, str]]]] = [
-    (
-        "Spec4",
-        [
-            ("How It Works", "https://spec4.ai/how-it-works/"),
-            ("About", "https://spec4.ai/about/"),
-            ("Get Started", "https://github.com/robertcrowe/Spec4"),
-        ],
-    ),
-    (
-        "Agents",
-        [
-            ("Reviewer", "https://spec4.ai/agents/reviewer/"),
-            ("Brainstormer", "https://spec4.ai/agents/brainstormer/"),
-            ("StackAdvisor", "https://spec4.ai/agents/stackadvisor/"),
-            ("Phaser", "https://spec4.ai/agents/phaser/"),
-        ],
-    ),
-    (
-        "Resources",
-        [
-            ("Community", "https://github.com/robertcrowe/Spec4/discussions"),
-            ("GitHub", "https://github.com/robertcrowe/Spec4"),
-            (
-                "Contributing",
-                "https://github.com/robertcrowe/Spec4/blob/main/CONTRIBUTING.md",
-            ),
-        ],
-    ),
+__all__ = [
+    "_card",
+    "_error",
+    "_feature_card",
+    "_footer",
+    "_nav_drawer",
+    "_render_message",
+    "_reformat_inline_lists",
+    "_agent_status_bar",
+    "_chat_action_buttons",
+    "_chat_layout",
+    "_done_layout",
+    "_setup_layout",
+    "_landing_layout",
+    "_working_dir_layout",
+    "_agent_select_layout",
 ]
 
 
-def _nav_drawer() -> html.Div:
-    def _link(text: str, href: str) -> Any:
-        return html.A(
-            text, href=href, target="_blank", rel="noopener", className="footer-link"
-        )
-
-    sections = [
-        html.Div(
-            [html.Div(heading, className="footer-col-heading")]
-            + [_link(label, url) for label, url in links],
-            style={"marginBottom": "1.5rem"},
-        )
-        for heading, links in _NAV_LINKS
-    ]
-    # Remove marginBottom from the last section
-    sections[-1] = html.Div(
-        [html.Div(_NAV_LINKS[-1][0], className="footer-col-heading")]
-        + [_link(label, url) for label, url in _NAV_LINKS[-1][1]],
-    )
-
-    return html.Div(
-        [
-            html.Div(id="nav-overlay", className="nav-overlay"),
-            html.Div(
-                id="nav-drawer",
-                className="nav-drawer",
-                children=[
-                    html.Button("✕", id="nav-close-btn", className="nav-close-btn"),
-                    html.Div(style={"clear": "both", "marginBottom": "1.5rem"}),
-                    *sections,
-                ],
-            ),
-        ]
-    )
-
-
-def _footer() -> html.Footer:
-    return html.Footer(
-        [
-            dmc.SimpleGrid(
-                cols={"base": 1, "sm": 2},
-                spacing="xl",
-                mb="lg",
-                children=[
-                    html.Div(
-                        [
-                            html.Div(
-                                [
-                                    html.Span("Spec", className="logo-spec"),
-                                    html.Span("4", className="logo-4"),
-                                    html.Span(" AI", className="logo-spec"),
-                                ],
-                                className="footer-brand-logo",
-                            ),
-                            html.P(
-                                "Spec-driven development, from idea to implementation plan.",  # noqa: E501
-                                className="footer-brand-tagline",
-                            ),
-                        ]
-                    ),
-                    dmc.SimpleGrid(
-                        cols=3,
-                        spacing="md",
-                        children=[
-                            html.Div(
-                                [html.Div(heading, className="footer-col-heading")]
-                                + [
-                                    html.A(
-                                        label,
-                                        href=url,
-                                        target="_blank",
-                                        rel="noopener",
-                                        className="footer-link",
-                                    )
-                                    for label, url in links
-                                ],
-                            )
-                            for heading, links in _NAV_LINKS
-                        ],
-                    ),
-                ],
-            ),
-            html.Div(
-                "© 2026 Robert Crowe. Open source under Apache 2.0 License.",
-                className="footer-bottom",
-            ),
-        ],
-        className="footer",
-    )
-
-
 # ---------------------------------------------------------------------------
-# Phase layouts
+# Landing
 # ---------------------------------------------------------------------------
 
 
@@ -332,6 +206,11 @@ def _landing_layout() -> html.Div:
     )
 
 
+# ---------------------------------------------------------------------------
+# Working directory browser
+# ---------------------------------------------------------------------------
+
+
 def _working_dir_layout(session: dict[str, Any]) -> html.Div:
     browser_path = session.get("browser_path") or str(pathlib.Path.home())
     current = pathlib.Path(browser_path)
@@ -436,213 +315,9 @@ def _working_dir_layout(session: dict[str, Any]) -> html.Div:
     )
 
 
-def _setup_provider_layout(
-    session: dict[str, Any],
-    prefs: dict[str, Any],
-    labels: list[str],
-    setup_error: str | None,
-) -> html.Div:
-    saved_prov = prefs.get("provider")
-    default_label = (
-        providers.PROVIDERS[saved_prov]["label"]
-        if saved_prov in providers.PROVIDERS
-        else labels[0]
-    )
-    return html.Div(
-        [
-            dmc.Title("Connect to an LLM provider", order=3, mb="sm"),
-            dmc.Text(
-                "Spec4 works with a wide variety of LLM providers and models. "
-                "Choose the one that works best for you.",
-                c="dimmed",
-                mb="sm",
-            ),
-            dmc.Alert(
-                "Note: Your API key is never stored outside of your system.",
-                color="blue",
-                variant="light",
-                mb="lg",
-            ),
-            _card(
-                dmc.Select(
-                    id="setup-provider",
-                    label="Provider",
-                    data=labels,
-                    value=default_label,
-                    mb="md",
-                ),
-                dmc.PasswordInput(
-                    id="setup-api-key",
-                    label="API Key",
-                    value=prefs.get("api_key") or "",
-                    mb="md",
-                ),
-                dmc.Checkbox(
-                    id="setup-save-prefs",
-                    label="Remember provider and keys in this browser? (stored in localStorage only)",  # noqa: E501
-                    checked=bool(prefs.get("save_prefs")),
-                    mb="md",
-                ),
-                _error(setup_error) if setup_error else html.Div(),
-                dmc.Group(
-                    [
-                        dmc.Button(
-                            "← Back",
-                            id="btn-setup-back-to-dir",
-                            variant="outline",
-                            color="gray",
-                        ),
-                        dmc.Button("Connect", id="btn-setup-connect"),
-                        dmc.Button(
-                            "Clear saved credentials",
-                            id="btn-setup-clear",
-                            variant="outline",
-                            color="red",
-                            disabled=not bool(prefs),
-                        ),
-                    ],
-                    mt="sm",
-                ),
-            ),
-        ]
-    )
-
-
-def _setup_model_layout(
-    session: dict[str, Any],
-    prefs: dict[str, Any],
-    setup_error: str | None,
-) -> html.Div:
-    available = session["available_models"]
-    saved_model = prefs.get("model")
-    default_model = (
-        saved_model
-        if saved_model in available
-        else (available[0] if available else None)
-    )
-    provider_label = providers.PROVIDERS[session["provider"]]["label"]
-    return html.Div(
-        [
-            dmc.Title("Select a Model", order=3, mb="sm"),
-            dmc.Text(
-                "Now that you have a provider, please select one of the models that this provider provides. "  # noqa: E501
-                "Remember that different models have different capabilities and different costs.",  # noqa: E501
-                c="dimmed",
-                mb="lg",
-            ),
-            _card(
-                dmc.Alert(f"Connected to {provider_label}", color="green", mb="md"),
-                dmc.Select(
-                    id="setup-model",
-                    label="Model",
-                    data=available,
-                    value=default_model,
-                    mb="md",
-                ),
-                _error(setup_error) if setup_error else html.Div(),
-                dmc.Group(
-                    [
-                        dmc.Button(
-                            "← Change Provider",
-                            id="btn-setup-back-provider",
-                            variant="outline",
-                            color="gray",
-                        ),
-                        dmc.Button("Continue →", id="btn-setup-model-continue"),
-                    ],
-                    mt="sm",
-                ),
-            ),
-        ]
-    )
-
-
-def _setup_tavily_layout(
-    session: dict[str, Any],
-    prefs: dict[str, Any],
-    setup_error: str | None,
-    image_support: bool | None = None,
-) -> html.Div:
-    if image_support is True:
-        image_alert: Any = dmc.Alert(
-            "This model supports image input — screenshot examples are available "
-            "in the Designer step.",
-            title="Image Support",
-            color="green",
-            variant="light",
-            mb="md",
-        )
-    elif image_support is False:
-        image_alert = dmc.Alert(
-            "This model does not support image input — screenshot upload will be "
-            "disabled in the Designer step. Go back to choose a different model if "
-            "you need image support.",
-            title="No Image Support",
-            color="orange",
-            variant="light",
-            mb="md",
-        )
-    else:
-        image_alert = html.Div()
-    return html.Div(
-        [
-            dmc.Title("Connect to Tavily Web Search", order=3, mb="sm"),
-            dmc.Text(
-                "Web search tends to be fairly useful when you're creating a spec for an application. "  # noqa: E501
-                "It will allow Spec4 to do things like look up the features of a library you might want "  # noqa: E501
-                "to use, or compare and contrast two different protocols that you're considering.",  # noqa: E501
-                c="dimmed",
-                mb="lg",
-            ),
-            _card(
-                dmc.Alert(f"LLM: {session['model']}", color="green", mb="md"),
-                image_alert,
-                dmc.Text(
-                    "Enables all agents to search the web for current information. "
-                    "Optional — skip if you don't have a Tavily key.",
-                    c="dimmed",
-                    mb="md",
-                ),
-                dmc.PasswordInput(
-                    id="setup-tavily-key",
-                    label="Tavily API Key",
-                    placeholder="tvly-…",
-                    value=prefs.get("tavily_key") or "",
-                    mb="md",
-                ),
-                _error(setup_error) if setup_error else html.Div(),
-                dmc.Group(
-                    [
-                        dmc.Button(
-                            "← Change Model",
-                            id="btn-setup-back-model",
-                            variant="outline",
-                            color="gray",
-                        ),
-                        dmc.Button(
-                            "Skip →", id="btn-setup-tavily-skip", variant="outline"
-                        ),
-                        dmc.Button("Connect & Start →", id="btn-setup-tavily-connect"),
-                    ],
-                    mt="sm",
-                ),
-            ),
-        ]
-    )
-
-
-def _setup_layout(
-    session: dict[str, Any],
-    prefs: dict[str, Any],
-    image_support: bool | None = None,
-) -> html.Div:
-    labels = providers.all_provider_labels()
-    setup_error = session.get("setup_error")
-    if session.get("available_models") is None:
-        return _setup_provider_layout(session, prefs, labels, setup_error)
-    if session.get("model") is None:
-        return _setup_model_layout(session, prefs, setup_error)
-    return _setup_tavily_layout(session, prefs, setup_error, image_support)
+# ---------------------------------------------------------------------------
+# Agent select
+# ---------------------------------------------------------------------------
 
 
 def _agent_select_layout(session: dict[str, Any]) -> html.Div:
@@ -831,318 +506,3 @@ def _agent_select_layout(session: dict[str, Any]) -> html.Div:
     )
 
     return html.Div(children)
-
-
-def _agent_status_bar(session: dict[str, Any]) -> html.Div:
-    active = session.get("active_agent", "brainstormer")
-    agents = [
-        ("reviewer", "🔍 Reviewer", session.get("code_review") is not None),
-        (
-            "brainstormer",
-            "🧠 Brainstormer",
-            session.get("vision_statement") is not None,
-        ),
-        ("stack_advisor", "⚙️ StackAdvisor", session.get("stack_statement") is not None),
-        ("phaser", "📋 Phaser", session.get("phaser_state") == STATE_PHASES_COMPLETE),
-    ]
-    items = []
-    for i, (key, label, done) in enumerate(agents):
-        if key == active:
-            items.append(dmc.Badge(label, color="green", variant="filled", size="md"))
-        elif done:
-            items.append(
-                dmc.Badge(f"✓ {label}", color="gray", variant="light", size="md")
-            )
-        else:
-            items.append(dmc.Badge(label, color="gray", variant="outline", size="md"))
-        if i < len(agents) - 1:
-            items.append(dmc.Text("→", c="dimmed", size="sm"))
-    return html.Div(
-        [
-            dmc.Group(
-                [
-                    dmc.Group(items, gap="xs"),
-                    dmc.Button(
-                        "← Back",
-                        id="btn-chat-back",
-                        variant="filled",
-                        size="xs",
-                        color="blue",
-                    ),
-                ],
-                justify="space-between",
-                mb="sm",
-            ),
-            dmc.Divider(mb="md"),
-        ]
-    )
-
-
-def _chat_action_buttons(session: dict[str, Any]) -> html.Div:
-    active = session.get("active_agent")
-    buttons = []
-
-    if active == "reviewer" and session.get("reviewer_state") == STATE_REVIEW_COMPLETE:
-        buttons = [
-            dmc.Button(
-                "💾 Download code_review.json", id="btn-dl-review", variant="outline"
-            ),
-            dmc.Button("Continue to Brainstormer →", id="btn-review-to-brainstormer"),
-        ]
-    elif (
-        active == "brainstormer"
-        and session.get("brainstormer_state") == STATE_VISION_COMPLETE
-    ):
-        buttons = [
-            dmc.Button(
-                "💾 Download vision.json", id="btn-dl-vision", variant="outline"
-            ),
-            dmc.Button("Continue to Designer →", id="btn-brainstormer-to-designer"),
-        ]
-    elif active == "stack_advisor":
-        back = dmc.Button(
-            "← Back to Designer",
-            id="btn-stack-to-brainstormer",
-            variant="outline",
-            color="gray",
-        )
-        if session.get("stack_advisor_state") == STATE_STACK_COMPLETE:
-            buttons = [
-                back,
-                dmc.Button(
-                    "💾 Download stack.json", id="btn-dl-stack", variant="outline"
-                ),
-                dmc.Button("Send to Phaser →", id="btn-stack-to-phaser"),
-            ]
-        else:
-            buttons = [back]
-    elif active == "phaser":
-        back = dmc.Button(
-            "← Back to Stack Advisor",
-            id="btn-phaser-to-stack",
-            variant="outline",
-            color="gray",
-        )
-        if session.get("phases"):
-            buttons = [
-                back,
-                dmc.Button(
-                    "💾 Download phases.zip", id="btn-dl-phases", variant="outline"
-                ),
-                dmc.Button("Done →", id="btn-phaser-done"),
-            ]
-        else:
-            buttons = [back]
-
-    if not buttons:
-        return html.Div()
-    return html.Div(
-        [
-            dmc.Divider(mb="sm"),
-            dmc.Group(buttons, mb="md"),
-        ]
-    )
-
-
-def _reformat_inline_lists(text: str) -> str:
-    """Break inline numbered lists onto separate lines for proper Markdown rendering."""
-    # Insert a newline before each "N. " that isn't already at the start of a line
-    text = re.sub(r"(?<!\n)[ \t]+(\d+)\.[ \t]+", r"\n\1. ", text)
-    # Ensure a blank line separates the preamble from the first list item
-    text = re.sub(r"([^\n])\n(1\. )", r"\1\n\n\2", text)
-    return text
-
-
-def _render_message(msg: dict[str, Any]) -> html.Div:
-    is_user = msg["role"] == "user"
-    content = msg["content"] if is_user else _reformat_inline_lists(msg["content"])
-    return html.Div(
-        dmc.Paper(
-            dcc.Markdown(content, style={"margin": 0}),
-            p="sm",
-            radius="md",
-            className="chat-bubble-user" if is_user else "chat-bubble-assistant",
-            style={"maxWidth": "85%"} if is_user else {"width": "100%"},
-        ),
-        style={
-            "display": "flex",
-            "justifyContent": "flex-end" if is_user else "flex-start",
-            "marginBottom": "8px",
-        },
-    )
-
-
-def _chat_layout(session: dict[str, Any]) -> html.Div:
-    messages = session.get("messages", [])
-    needs_init = not messages and not session.get("_initial_turn_done")
-
-    return html.Div(
-        [
-            # Trigger initial agent turn once on first render.
-            # max_intervals=0 disables the interval (never fires) when not needed,
-            # but keeps n_intervals available as a callback input.
-            dcc.Interval(
-                id="init-turn-interval",
-                interval=300,
-                max_intervals=1 if needs_init else 0,
-            ),
-            # Always-present download triggers (invisible)
-            dcc.Download(id="dl-vision"),
-            dcc.Download(id="dl-stack"),
-            dcc.Download(id="dl-code-review"),
-            dcc.Download(id="dl-phases"),
-            _agent_status_bar(session),
-            html.Div(
-                html.Div(
-                    [_render_message(m) for m in messages]
-                    + (
-                        [dmc.Text("Thinking…", c="dimmed", size="sm")]
-                        if needs_init
-                        else []
-                    ),
-                    style={"display": "flex", "flexDirection": "column"},
-                ),
-                id="chat-scroll-area",
-                style={
-                    "height": "450px",
-                    "overflowY": "auto",
-                    "marginBottom": "var(--mantine-spacing-md)",
-                },
-            ),
-            _chat_action_buttons(session),
-            html.Div(
-                dmc.Progress(
-                    value=100, animated=True, striped=True, color="blue", size="sm"
-                ),
-                id="chat-progress-container",
-                style={"display": "none", "marginBottom": "12px"},
-            ),
-            html.Div(
-                [
-                    dmc.Textarea(
-                        id="chat-input",
-                        placeholder="Type your message…",
-                        style={"flex": "1"},
-                        autosize=True,
-                        minRows=2,
-                        n_submit=0,
-                    ),
-                    dmc.Button("Send", id="btn-chat-submit"),
-                ],
-                style={
-                    "display": "flex",
-                    "alignItems": "stretch",
-                    "gap": "var(--mantine-spacing-sm)",
-                    "width": "100%",
-                },
-            ),
-        ]
-    )
-
-
-def _done_layout(session: dict[str, Any]) -> html.Div:
-    phases = session.get("phases", [])
-    n = len(phases)
-    spec4_dir = pathlib.Path(session.get("working_dir", "")) / ".spec4"
-    has_vision = (spec4_dir / "vision.json").exists()
-    has_stack = (spec4_dir / "stack.json").exists()
-
-    phase_rows = [
-        dmc.Text(
-            f"Phase {p.get('phase_number')}: {p.get('phase_title', '')}",
-            size="sm",
-            style={
-                "paddingLeft": "0.5rem",
-                "borderLeft": "2px solid var(--mantine-color-blue-5)",
-            },
-            mb="xs",
-        )
-        for p in sorted(phases, key=lambda x: x.get("phase_number", 0))
-    ]
-
-    return html.Div(
-        [
-            dcc.Download(id="dl-phases-done"),
-            dcc.Download(id="dl-vision-done"),
-            dcc.Download(id="dl-stack-done"),
-            dmc.Title("Your Project Phases Are Ready", order=3, mb="xs"),
-            dmc.Text(
-                f"Spec4 has broken your project into {n} executable phase{'s' if n != 1 else ''}. "  # noqa: E501
-                "Hand them to an agentic coding tool and work through them in order.",
-                c="dimmed",
-                mb="lg",
-            ),
-            _card(
-                dmc.Title("How to use your phases", order=5, mb="sm"),
-                dcc.Markdown(
-                    "1. Open your agentic coding tool (Claude Code, Codex, Antigravity, Hermes, etc.) "  # noqa: E501
-                    "inside your project directory.\n"
-                    "2. Load or paste the contents of **Phase 1** as your prompt.\n"
-                    "3. Let the agent complete the phase fully — verify it passes before continuing.\n"  # noqa: E501
-                    "4. Repeat for Phase 2, Phase 3, and so on **in order**.\n\n"
-                    "> Do not skip phases. Each phase builds directly on the one before it.\n\n"  # noqa: E501
-                    "**Loading options:**\n"
-                    "- **Copy/paste** — open each `.json` file and paste the contents into your tool.\n"  # noqa: E501
-                    "- **File loading** — tools like Claude Code accept files directly "
-                    "(e.g. `Please implement @.spec4/phases/phase1.json`).\n"
-                    "- **Download** — use the button below to get all phases as a zip.",
-                    style={"fontSize": "0.9rem"},
-                ),
-                mb="md",
-            ),
-            _card(
-                dmc.Title(
-                    "Python project? Set up your environment first", order=5, mb="sm"
-                ),
-                dcc.Markdown(
-                    "Before starting your agentic coding tool, initialise a virtual environment "  # noqa: E501
-                    "so your agent installs dependencies in an isolated, reproducible way:\n\n"  # noqa: E501
-                    "```bash\n"
-                    "uv init\n"
-                    "uv sync\n"
-                    "source .venv/bin/activate\n"
-                    "```\n\n"
-                    "On Windows use `.venv\\Scripts\\activate` instead of `source .venv/bin/activate`.",  # noqa: E501
-                    style={"fontSize": "0.9rem"},
-                ),
-                mb="md",
-            ),
-            _card(
-                dmc.Title(f"Your {n} phase{'s' if n != 1 else ''}", order=5, mb="sm"),
-                *phase_rows,
-            )
-            if phase_rows
-            else html.Div(),
-            dmc.Divider(my="md"),
-            dmc.Group(
-                [
-                    dmc.Button(
-                        "← Back to Phaser",
-                        id="btn-done-back-to-phaser",
-                        variant="outline",
-                        color="gray",
-                    ),
-                    dmc.Button(
-                        "💾 Download vision.json",
-                        id="btn-dl-vision-done",
-                        variant="outline",
-                        style={"display": "none"} if not has_vision else {},
-                    ),
-                    dmc.Button(
-                        "💾 Download stack.json",
-                        id="btn-dl-stack-done",
-                        variant="outline",
-                        style={"display": "none"} if not has_stack else {},
-                    ),
-                    dmc.Button(
-                        "💾 Download phases.zip",
-                        id="btn-dl-phases-done",
-                        variant="outline",
-                    ),
-                    dmc.Button(
-                        "Start New Project", id="btn-done-new-project", variant="light"
-                    ),
-                ]
-            ),
-        ]
-    )
